moved the IDs to uint8_t
svn path=/trunk/iCub/firmware/emBODY/any-arch/sys/embobj/; revision=24981

+++ emBODY/eBcode/arch-arm/embobj/wip-icub/api/EoBoards.h
@@ -54,9 +54,7 @@ extern "C" {
 
 // - declaration of public user-defined types ------------------------------------------------------------------------- 
 
-typedef uint16_t eObrd_boardId_t;
-
-#warning --> eval to use uint8_t for eObrd_boardId_t ...
+typedef uint8_t eObrd_boardId_t;
 
 
 typedef enum
